Update to shim 15.1
This upstream tag does not build, so I added a small patch in
`build-fix.patch` that gets applied after cloning the repo.

Note that we are still building with a Fedora 28 container. I tried
updating to newer versions, but the shim build fails on all newer
versions I tried. Some of these errors have likely been fixed in the
past two years of shim development, but there haven't been any
official releases of that code.

https://neverware.atlassian.net/browse/OVER-12692